pyproject.toml
[tool.poetry]
name = "quiz-bot"
version = "0.1.0"
description = ""
authors = ["Tatiana_Garbuzova <tatiannna0697@gmail.com>"]
readme = "README.md"
packages = [{include = "quiz_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.13.0"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
bot = "quiz_bot.main:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
